+++ src/unitycrunch.rs
@@ -29,8 +29,8 @@ fn unity_crunch_unpack_level(
         };
     let width = core::cmp::max(1, tex_info.width >> level_index);
     let height = core::cmp::max(1, tex_info.height >> level_index);
-    let blocks_x: u32 = core::cmp::max(1, ((width + 3) >> 2) as u32);
-    let blocks_y: u32 = core::cmp::max(1, ((height + 3) >> 2) as u32);
+    let blocks_x: u32 = core::cmp::max(1, (width + 3) >> 2);
+    let blocks_y: u32 = core::cmp::max(1, (height + 3) >> 2);
     let row_pitch: u32 = blocks_x
         * match crn_decomp::crnd_get_bytes_per_dxt_block(&mut tex_info.format) {
             Ok(s) => s,
